integrate queryUpdateAction into updateAction
