+++ lib/models/actions.js
@@ -34,7 +34,7 @@ var util = require('util'),
     updateAction = require('./updateAction'),
     postAction = require('./postAction'),
     twitterAction = require('./twitterAction'),
-    queryUpdateAction = require('./queryUpdateAction'),
+    //queryUpdateAction = require('./queryUpdateAction'),
     errors = {},
     myutils = require('../myutils'),
     logger = require('logops'),
@@ -71,7 +71,7 @@ function validateAction(axn) {
             case 'update':
             case 'post':
             case 'twitter':
-            case 'queryUpdate':
+                //        case 'queryUpdate':
                 break;
             default:
                 return new errors.UnknownActionType(axnElem.type);
@@ -117,9 +117,9 @@ function conditionalExec(task, lastTime, callbackW) {
                         case 'update':
                             updateAction.doIt(action, event, callbackS);
                             break;
-                        case 'queryUpdate':
-                            queryUpdateAction.doIt(action, event, callbackS);
-                            break;
+                        // case 'queryUpdate':
+                        //     queryUpdateAction.doIt(action, event, callbackS);
+                        //     break;
                         case 'post':
                             postAction.doIt(action, event, callbackS);
                             break;
