+++ lib/models/updateAction.js
@@ -119,6 +119,21 @@ function buildUpdateOptions(action, event) {
 
     return options;
 }
+
+function buildQueryOptionsParams(action, event) {
+    var filter;
+
+    // Options detail:
+    // https://conwetlab.github.io/ngsijs/stable/NGSI.Connection.html#.%22v2.listEntities%22__anchor
+
+    action.parameters.filter = action.parameters.filter || '${filter}';
+
+    filter = myutils.expandVar(action.parameters.filter, event);
+
+    logger.debug('filter built: %j', filter);
+    return filter;
+}
+
 function doRequest(action, event, token, callback) {
     var subservice, service, headers, options, updateOrion;
 
@@ -297,6 +312,65 @@ function processOptionParams(action, event) {
 }
 
 /**
+ * Process NGSIv2 action do request option Params
+ *
+ * @param action The request update action information
+ * @param entity NGSI entity
+ *
+ * @return options object
+ */
+function processUpdateOptionParams(action, entity) {
+    action.parameters.attributes = action.parameters.attributes || [];
+
+    var options = {};
+    action.parameters.attributes.forEach(function(attr) {
+        // Direct value for Text, DateTime, and others. Apply expandVar for strings
+        let theValue = attr.name in entity ? entity[attr.name].value : attr.value;
+        let theType = attr.name in entity ? entity[attr.name].type : attr.type;
+        // Metadata should be null or object
+        let theMeta = attr.metadata;
+        var date;
+
+        switch (theType) {
+            case 'Text':
+                theValue = theValue.toString();
+                break;
+            case 'Number':
+                theValue = parseFloat(theValue);
+                break;
+            case 'Boolean':
+                if (typeof theValue === 'string') {
+                    theValue = theValue.toLowerCase().trim() === 'true';
+                }
+                break;
+            case 'DateTime':
+                if (parseInt(theValue).toString() === theValue) {
+                    // Parse String with number (timestamp__ts in Perseo events)
+                    theValue = parseInt(theValue);
+                }
+                date = new Date(theValue);
+                theValue = isNaN(date.getTime()) ? theValue : date.toISOString();
+                break;
+            case 'None':
+                theValue = null;
+                break;
+        }
+        var key = attr.name;
+        options[key] = {
+            value: theValue,
+            type: theType
+        };
+        if (attr.metadata !== undefined) {
+            options[key].metadata = theMeta;
+        }
+    });
+    options.id = entity.id;
+    options.type = entity.type;
+    logger.debug('options built: %j', options);
+    return options;
+}
+
+/**
  * Manage update action request for NGSv2 request
  *
  * @param action The request update action information
@@ -316,21 +390,98 @@ function doRequestV2(action, event, token, callback) {
     }
     var connection = new NGSI.Connection(config.orion.URL, options);
 
-    var changes = processOptionParams(action, event);
-    metrics.IncMetrics(event.service, event.subservice, metrics.actionEntityUpdate);
+    // check if filter in action
+    if (action && action.parameters && action.parameter.filter) {
+        // UpdateAction with filter
+        logger.debug('action: %j, event: %j', action, event);
 
-    connection.v2.createEntity(changes, { upsert: true }).then(
-        (response) => {
-            metrics.IncMetrics(event.service, event.subservice, metrics.okActionEntityUpdate);
-            alarm.release(alarm.ORION);
-            callback(null, response);
-        },
-        (error) => {
-            metrics.IncMetrics(event.service, event.subservice, metrics.failedActionEntityUpdate);
-            alarm.raise(alarm.ORION, null, error);
-            callback(error, null);
-        }
-    );
+        var queryOptions = buildQueryOptionsParams(action, event);
+
+        //
+        // 1. Make query with filter
+        //
+        connection.v2.listEntities(queryOptions).then(
+            (response) => {
+                // Entities retrieved successfully
+                // response.results is an array with the retrieved entities
+                // response.correlator transaction id associated with the server response
+                // response.count contains the total number of entities selected
+                //   by this query
+                // response.offset contains the offset used in the request
+
+                metrics.IncMetrics(event.service, event.subservice, metrics.actionEntityFilter);
+                alarm.release(alarm.ORION);
+
+                logger.debug('full query response: %j', response);
+
+                var changes = {
+                    actionType: 'APPEND',
+                    entities: []
+                };
+
+                if (options.actionType !== undefined) {
+                    changes.actionType.updateAction = options.actionType;
+                }
+
+                // Process response (array of entities)
+                response.results.forEach(function(entity) {
+                    var updateOptions = processUpdateOptionParams(action, entity);
+                    changes.entities.push(updateOptions);
+                }); // For each
+
+                logger.debug('entities to update: %j', changes);
+                metrics.IncMetrics(event.service, event.subservice, metrics.actionEntityUpdate);
+                //
+                // 2. Apply update to all entities in a batch
+                //
+                connection.v2.batchUpdate(changes, { upsert: true }).then(
+                    (response) => {
+                        // Entities updated successfully
+                        // response.correlator transaction id associated with the server response
+                        logger.debug('response batchUpdate: %j', response);
+                        metrics.IncMetrics(event.service, event.subservice, metrics.okActionEntityUpdate);
+                        alarm.release(alarm.ORION);
+                        callback(null, response);
+                    },
+                    (error) => {
+                        // Error updating entities
+                        // If the error was reported by Orion, error.correlator will be
+                        // filled with the associated transaction id
+                        metrics.IncMetrics(event.service, event.subservice, metrics.failedActionEntityUpdate);
+                        logger.warn('error batchUpdate%j', error);
+                        alarm.raise(alarm.ORION, null, error);
+                    }
+                ); // batchUpdate
+            },
+            (error) => {
+                // Error retrieving entities
+                // If the error was reported by Orion, error.correlator will be
+                // filled with the associated transaction id
+                metrics.IncMetrics(event.service, event.subservice, metrics.failedActionEntityUpdate);
+                logger.warn('error filtering %j', error);
+                alarm.raise(alarm.ORION, null, error);
+                callback(error, null);
+            }
+        );
+    } else {
+        // classical updateAction (without filter)
+
+        var changes = processOptionParams(action, event);
+        metrics.IncMetrics(event.service, event.subservice, metrics.actionEntityUpdate);
+
+        connection.v2.createEntity(changes, { upsert: true }).then(
+            (response) => {
+                metrics.IncMetrics(event.service, event.subservice, metrics.okActionEntityUpdate);
+                alarm.release(alarm.ORION);
+                callback(null, response);
+            },
+            (error) => {
+                metrics.IncMetrics(event.service, event.subservice, metrics.failedActionEntityUpdate);
+                alarm.raise(alarm.ORION, null, error);
+                callback(error, null);
+            }
+        );
+    }
 }
 
 function doItWithTokenV2(action, event, callback) {
